move adjtime
Storing adjtime on /var is causing too much trouble as it may not be mounted at early boot.

This reverts to upstream behavior and allows us to simplify initscripts.

This means the hwclock 'daemon' will write to /etc. If this is an issue it should be replaced
by something better such as the chrony ntp daemon, which handles the RTC in a sensible way.